Migrate NetcdfEncoder tests to munit
